+++ packages/client-flows/src/lib/config.ts
@@ -61,7 +61,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "0.0.0",
-    sdkVersion: "1.0.0-alpha.19",
-    genVersion: "2.393.4",
-    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.19 2.393.4 0.0.0 @formance/sdk-flows",
+    sdkVersion: "1.0.0-alpha.20",
+    genVersion: "2.396.0",
+    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.20 2.396.0 0.0.0 @formance/sdk-flows",
 } as const;