+++ packages/client-flows/RELEASES.md
@@ -199,3 +199,13 @@ Based on:
 - [typescript v1.0.0-alpha.19] packages/client-flows
 ### Releases
 - [NPM v1.0.0-alpha.19] https://www.npmjs.com/package/@formance/sdk-flows/v/1.0.0-alpha.19 - packages/client-flows
+
+## 2024-08-13 00:12:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.363.1 (2.396.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.20] packages/client-flows
+### Releases
+- [NPM v1.0.0-alpha.20] https://www.npmjs.com/package/@formance/sdk-flows/v/1.0.0-alpha.20 - packages/client-flows